requirements.txt
Flask==2.3.2
pymongo==4.5.0
python-dotenv==1.0.0
cloudinary==1.36.0
gunicorn==21.2.0
